# docs/swagger.yaml
basePath: /api
definitions:
  controllers.APIResponse:
    properties:
      msg:
        type: string
    type: object
  controllers.APIUserResponse:
    properties:
      user:
        $ref: '#/definitions/entities.User'
    type: object
...
  entities.JSONResponse:
    properties:
      data: {}
      error:
        type: boolean
      message:
        type: string
    type: object
...
  entities.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      isVender:
        type: string
      password:
        type: string
      password_inserted_at:
        type: string
      password_reset_token:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
  entities.UserPayload:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      is_vendor:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
info:
  contact:
    email: bicosteve4@gmail.com
    name: Bico Oloo
    url: https://github.com/bicosteve
  description: Booking API to make booking reservations
  title: Booking API
  version: "1.0"
paths:
...
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Receives user payload, validate it then send it to service
      operationId: login-user
      parameters:
      - description: Login User
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/entities.UserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: '{"token":"xxxxxxxxxxx"}'
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad Request, validation error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "404":
          description: Bad Request, user not found
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: Authorize User
      tags:
      - auth
  /api/user/me:
    get:
      description: Returns logged in user details
      operationId: user-profile
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            $ref: '#/definitions/controllers.APIUserResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: Get a  User
      tags:
      - auth
  /api/user/password-reset:
    put:
      consumes:
      - application/json
      description: Receives user payload, validate it then send it to service
      operationId: reset-password
      parameters:
      - description: Generate auth token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/entities.UserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Returns user
          schema:
            $ref: '#/definitions/controllers.APIUserResponse'
        "400":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.APIUserResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.APIUserResponse'
      summary: Reset Password
      tags:
      - auth
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Receives user payload, validate it then send it to service
      operationId: register-user
      parameters:
      - description: Register User
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/entities.UserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: User registered
          schema:
            $ref: '#/definitions/controllers.APIResponse'
        "400":
          description: Bad request, validation error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/entities.JSONResponse'
      summary: Registers User
      tags:
      - auth
  /api/user/reset:
    post:
      consumes:
      - application/json
      description: Receives user payload, validate it then send it to service
      operationId: reset-token
      parameters:
      - description: Generate auth token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/entities.UserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Returns user
          schema:
            $ref: '#/definitions/controllers.APIUserResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.APIUserResponse'
      summary: Generate Password Reset Token
      tags:
      - auth
...
schemes:
- http
swagger: "2.0"